mybackend/requirements.txt
Django>=5.2,<6.0
psycopg2-binary>=2.9
 djangorestframework>=3.14
 django-cors-headers>=4.0
 djangorestframework-simplejwt>=5.2
 gunicorn>=21.2 